Updates test discovery step in build (#29)
CMake will run built gtest executables as part of the build in order to
determine the tests in the executable. The default timeout for this run
was 5 seconds but our CI system has been hitting this recently. In order
to accomedating slow VMs in the cloud we are extending the timeout to 60
seconds.