K4AViewer: fix crash when K4A mic is muted
When the OS mutes the K4A mic, libsoundio gives back a 'hole' in the sound buffer. We were treating this as an error condition, which it actually is not, and trying to fail. This change stops treating the 'buffer hole' as an error condition.

There was another bug where we weren't aborting the sound polling thread immediately upon realizing that the mic had failed, which was resulting in an AV when we tried to write back to the mic after tearing it down in the error handling code. This change also adds that missing abort.