Remove imported lib from libyuv (#129)
In libyuv the imported lib for the DLL has the same name as the static
lib. Thus, there is a target conflict that the lastest version of ninja
will complain about.

This works around this issue by tricking cmake into not generating a
import lib for libyuv as we use the static lib only.