RUNPATH only should be set during build
Currently, we set RUNPATH in all our shared objects and executables
at both build time and runtime. Build time is needed because we move
the binaries around to other machines. However, the install time is not
needed.

This is step 1 of 2 in order to move the depthengine into a private
folder instead of in the /lib folder