PythonK4A: Adding initial directory structure, k4atypes.py which defines enums, and tests to check the enum values.
The _bindings/ folder is meant to contain internal ctypes bindings to the C-exported functions in k4a.dll.
The _libs/ folder is meant to contain the required DLLs: k4a.dll and DepthEngine.dll.
The api/ folder is meant for modules that define the actual Python K4A API.
The docs/ folder is for documentation.
The tests/ folder is for unit tests and functional tests.
The examples/ folder is for simple example code on how to use the API.