SDK C API: Refactoring k4a.c and k4ainternal/modes.h so that the mode definitions are all stored in k4ainternal/modes.h. This consolidates all mode-related information in one file and removes 2 definitions of the same information in k4a.c and matroska_read.cpp. This also allows the convenience functions in k4ainternal/modes.h to be very much simplified.
UnitTests: Refactored the logic in writing and reading tags so that subsequent tags are not dependent on prior tags being successful. There was an issue in which calibration data was not being written successfully (because there is no device connected for unit tests), and because calibration data is one of the first tags written, all tags after it were skipped. This change allows tags to be written independently of the successful or failed writes of prior tags.